src/backend/requirements.txt
# FastAPI + Async
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
httpx==0.28.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
cloud-sql-python-connector[asyncpg]==1.14.0

# Auth
pyjwt[crypto]==2.10.1
bcrypt==4.2.1
passlib[bcrypt]==1.7.4

# Validation & Serialization
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Redis (Cloud Memorystore)
redis[hiredis]==5.2.1

# Celery (task queue for reminders)
celery[redis]==5.4.0

# Notifications (Infobip - Viber/WhatsApp/SMS)
infobip-api-python-sdk==4.0.0

# iCal parsing (Airbnb/Booking.com sync)
icalendar==6.1.0
recurring-ical-events==3.4.0

# e-Factura XML generation
lxml==5.3.0
xmlsec==1.3.14

# GCP
google-cloud-storage==2.19.0
google-cloud-tasks==2.17.0

# Utils
python-dateutil==2.9.0
pytz==2024.2
